pyproject.toml
[project]
name = "papermind"
version = "0.3.0"
description = "Automated literature survey — Main Agent + Multi Sub-Agent (MiniMax + DDG + FAISS RAG)"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pymupdf>=1.25.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.7.4",
    "ddgs>=9.0.0",
    "fastmcp>=2.0.0",
    "trafilatura>=1.12.0",
]

[project.scripts]
PM = "papermind.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/papermind",
    "src/shared",
    "src/agents",
    "src/mcp_servers",
    "src/orchestrator",
    "src/rag",
]
